Fixes some undefined issues in handleAddDownload
